Changed floats to doubles for better accuracy, fixed variables and math operations for seconds, and changed Math.ceil to Math.round for more accurate results for many items added.

+++ main/Playlist.java
@@ -10,7 +10,7 @@ public interface Playlist {
     /*
      * Calculates and returns the duration of the playlist
      */
-    float playlistDuration();
+    double playlistDuration();
     
     /*
      * Play next song, returning song info and removing it
